v003_018

+++ src/backup/train_v003_007.py
@@ -460,96 +460,105 @@ log_path.mkdir(parents=True, exist_ok=True)
 ####################################################################################################
 # Data Loading
 
-file_folder = '../input'
-train = pd.read_csv(f'{file_folder}/train.csv')
-test = pd.read_csv(f'{file_folder}/test.csv')
-sub = pd.read_csv(f'{file_folder}/sample_submission.csv')
-structures = pd.read_csv(f'{file_folder}/structures.csv')
-scalar_coupling_contributions = pd.read_csv(f'{file_folder}/scalar_coupling_contributions.csv')
-train_cos = unpickle("../processed/v001/train_003.df.pkl", )[["id", "f003:cos_0_1", "f003:cos_1"]]
-test_cos = unpickle("../processed/v001/test_003.df.pkl", )[["id", "f003:cos_0_1", "f003:cos_1"]]
 
-# train_angle_add = unpickle("../processed/v003/train_005.df.pkl", )
-# test_angle_add = unpickle("../processed/v003/test_005.df.pkl", )
+train_path = save_path / f"train_concat_xxx"
+test_path  = save_path / f"test_concat_xxx"
+if train_path.exists() and test_path.exists():
+    train = unpickle(train_path)
+    test = unpickle(test_path)
 
-train_add = unpickle("../processed/v003/train_006.df.pkl", )
-test_add = unpickle("../processed/v003/test_006.df.pkl", )
+else:
 
-babel_cols = ['id', 'Angle', 'Torsion', 'cos2T', 'cosT', 'sp']
-babel_train = pd.read_csv("../processed/v003/babel_train.csv", usecols=babel_cols)
-babel_test = pd.read_csv("../processed/v003/babel_test.csv", usecols=babel_cols)
+    file_folder = '../input'
+    train = pd.read_csv(f'{file_folder}/train.csv')
+    test = pd.read_csv(f'{file_folder}/test.csv')
+    sub = pd.read_csv(f'{file_folder}/sample_submission.csv')
+    structures = pd.read_csv(f'{file_folder}/structures.csv')
+    scalar_coupling_contributions = pd.read_csv(f'{file_folder}/scalar_coupling_contributions.csv')
+    train_cos = unpickle("../processed/v001/train_003.df.pkl", )[["id", "f003:cos_0_1", "f003:cos_1"]]
+    test_cos = unpickle("../processed/v001/test_003.df.pkl", )[["id", "f003:cos_0_1", "f003:cos_1"]]
 
-####################################################################################################
-# Feature Engineering
+    # train_angle_add = unpickle("../processed/v003/train_005.df.pkl", )
+    # test_angle_add = unpickle("../processed/v003/test_005.df.pkl", )
 
-train = pd.merge(train, scalar_coupling_contributions, how = 'left',
-                  left_on  = ['molecule_name', 'atom_index_0', 'atom_index_1', 'type'],
-                  right_on = ['molecule_name', 'atom_index_0', 'atom_index_1', 'type'])
+    train_add = unpickle("../processed/v003/train_006.df.pkl", )
+    test_add = unpickle("../processed/v003/test_006.df.pkl", )
 
-train = map_atom_info(train, 0)
-train = map_atom_info(train, 1)
-test = map_atom_info(test, 0)
-test = map_atom_info(test, 1)
+    babel_cols = ['id', 'Angle', 'Torsion', 'cos2T', 'cosT', 'sp']
+    babel_train = pd.read_csv("../processed/v003/babel_train.csv", usecols=babel_cols)
+    babel_test = pd.read_csv("../processed/v003/babel_test.csv", usecols=babel_cols)
 
-train_p_0 = train[['x_0', 'y_0', 'z_0']].values
-train_p_1 = train[['x_1', 'y_1', 'z_1']].values
-test_p_0 = test[['x_0', 'y_0', 'z_0']].values
-test_p_1 = test[['x_1', 'y_1', 'z_1']].values
+    ####################################################################################################
+    # Feature Engineering
 
-train['dist'] = np.linalg.norm(train_p_0 - train_p_1, axis=1)
-test['dist'] = np.linalg.norm(test_p_0 - test_p_1, axis=1)
-train['dist_x'] = (train['x_0'] - train['x_1']) ** 2
-test['dist_x'] = (test['x_0'] - test['x_1']) ** 2
-train['dist_y'] = (train['y_0'] - train['y_1']) ** 2
-test['dist_y'] = (test['y_0'] - test['y_1']) ** 2
-train['dist_z'] = (train['z_0'] - train['z_1']) ** 2
-test['dist_z'] = (test['z_0'] - test['z_1']) ** 2
+    train = pd.merge(train, scalar_coupling_contributions, how = 'left',
+                      left_on  = ['molecule_name', 'atom_index_0', 'atom_index_1', 'type'],
+                      right_on = ['molecule_name', 'atom_index_0', 'atom_index_1', 'type'])
 
-train['type_0'] = train['type'].apply(lambda x: x[0])
-test['type_0'] = test['type'].apply(lambda x: x[0])
+    train = map_atom_info(train, 0)
+    train = map_atom_info(train, 1)
+    test = map_atom_info(test, 0)
+    test = map_atom_info(test, 1)
 
-train['abs_dist'] = np.linalg.norm(train_p_0-train_p_1,axis=1,ord=1)
-test['abs_dist'] = np.linalg.norm(test_p_0-test_p_1,axis=1,ord=1)
-dist12('dist_xy','x','y')
-dist12('dist_xz','x','z')
-dist12('dist_yz','y','z')
+    train_p_0 = train[['x_0', 'y_0', 'z_0']].values
+    train_p_1 = train[['x_1', 'y_1', 'z_1']].values
+    test_p_0 = test[['x_0', 'y_0', 'z_0']].values
+    test_p_1 = test[['x_1', 'y_1', 'z_1']].values
 
-atom_count = structures.groupby(['molecule_name', 'atom']).size().unstack(fill_value=0)
-train = pd.merge(train, atom_count, how = 'left', left_on  = 'molecule_name', right_on = 'molecule_name')
-test = pd.merge(test, atom_count, how = 'left', left_on  = 'molecule_name', right_on = 'molecule_name')
+    train['dist'] = np.linalg.norm(train_p_0 - train_p_1, axis=1)
+    test['dist'] = np.linalg.norm(test_p_0 - test_p_1, axis=1)
+    train['dist_x'] = (train['x_0'] - train['x_1']) ** 2
+    test['dist_x'] = (test['x_0'] - test['x_1']) ** 2
+    train['dist_y'] = (train['y_0'] - train['y_1']) ** 2
+    test['dist_y'] = (test['y_0'] - test['y_1']) ** 2
+    train['dist_z'] = (train['z_0'] - train['z_1']) ** 2
+    test['dist_z'] = (test['z_0'] - test['z_1']) ** 2
 
-train = create_features(train)
-test = create_features(test)
+    train['type_0'] = train['type'].apply(lambda x: x[0])
+    test['type_0'] = test['type'].apply(lambda x: x[0])
 
-angle_df_train, angle_df_test = angle_feature_conv()
-train = train.merge(angle_df_train, on="id", how="left") # .merge(train_cos,  on="id", how="left")
-test = test.merge(angle_df_test, on="id", how="left") # .merge(test_cos,  on="id", how="left")
-# train = train.merge(train_angle_add, on="id", how="left")
-# test = test.merge(test_angle_add, on="id", how="left")
-train = train.merge(train_add, on="id", how="left")
-test = test.merge(test_add, on="id", how="left")
+    train['abs_dist'] = np.linalg.norm(train_p_0-train_p_1,axis=1,ord=1)
+    test['abs_dist'] = np.linalg.norm(test_p_0-test_p_1,axis=1,ord=1)
+    dist12('dist_xy','x','y')
+    dist12('dist_xz','x','z')
+    dist12('dist_yz','y','z')
 
-train = train.merge(babel_train, on="id", how="left")
-test = test.merge(babel_test, on="id", how="left")
+    atom_count = structures.groupby(['molecule_name', 'atom']).size().unstack(fill_value=0)
+    train = pd.merge(train, atom_count, how = 'left', left_on  = 'molecule_name', right_on = 'molecule_name')
+    test = pd.merge(test, atom_count, how = 'left', left_on  = 'molecule_name', right_on = 'molecule_name')
 
-ob_charges = pd.read_csv("../processed/v003/ob_charges.csv", index_col=0)
-train = map_ob_charges(train, 0)
-train = map_ob_charges(train, 1)
-test = map_ob_charges(test, 0)
-test = map_ob_charges(test, 1)
+    train = create_features(train)
+    test = create_features(test)
 
-train = reduce_mem_usage(train)
-test = reduce_mem_usage(test)
+    angle_df_train, angle_df_test = angle_feature_conv()
+    train = train.merge(angle_df_train, on="id", how="left") # .merge(train_cos,  on="id", how="left")
+    test = test.merge(angle_df_test, on="id", how="left") # .merge(test_cos,  on="id", how="left")
+    # train = train.merge(train_angle_add, on="id", how="left")
+    # test = test.merge(test_angle_add, on="id", how="left")
+    train = train.merge(train_add, on="id", how="left")
+    test = test.merge(test_add, on="id", how="left")
 
-for f in ['atom_1', 'type_0', 'type']:
-    if f in good_columns:
-        lbl = LabelEncoder()
-        lbl.fit(list(train[f].values) + list(test[f].values))
-        train[f] = lbl.transform(list(train[f].values))
-        test[f] = lbl.transform(list(test[f].values))
+    train = train.merge(babel_train, on="id", how="left")
+    test = test.merge(babel_test, on="id", how="left")
 
-to_pickle(save_path/f"train_concat_v003_{DATA_VERSION}_{TRIAL_NO}.pkl", train)
-to_pickle(save_path/f"test_concat_v003_{DATA_VERSION}_{TRIAL_NO}.pkl", test)
+    ob_charges = pd.read_csv("../processed/v003/ob_charges.csv", index_col=0)
+    train = map_ob_charges(train, 0)
+    train = map_ob_charges(train, 1)
+    test = map_ob_charges(test, 0)
+    test = map_ob_charges(test, 1)
+
+    train = reduce_mem_usage(train)
+    test = reduce_mem_usage(test)
+
+    for f in ['atom_1', 'type_0', 'type']:
+        if f in good_columns:
+            lbl = LabelEncoder()
+            lbl.fit(list(train[f].values) + list(test[f].values))
+            train[f] = lbl.transform(list(train[f].values))
+            test[f] = lbl.transform(list(test[f].values))
+
+    to_pickle(save_path/f"train_concat_v003_{DATA_VERSION}_{TRIAL_NO}.pkl", train)
+    to_pickle(save_path/f"test_concat_v003_{DATA_VERSION}_{TRIAL_NO}.pkl", test)
 
 X = train[good_columns].copy()
 y = train['scalar_coupling_constant']
@@ -564,39 +573,46 @@ pd.DataFrame({"columns": X.columns.tolist()}).to_csv(log_path/f"use_cols.csv")
 n_fold = 5
 folds = KFold(n_splits=n_fold, shuffle=True, random_state=11)
 
-params = {'num_leaves': 128,
-          'min_child_samples': 79,
-          'objective': 'regression',
-          'max_depth': 9,
-          'learning_rate': 0.2,
-          "boosting_type": "gbdt",
-          "subsample_freq": 1,
-          "subsample": 0.9,
-          "bagging_seed": 11,
-          "metric": 'mae',
-          "verbosity": -1,
-          'reg_alpha': 0.1,
-          'reg_lambda': 0.3,
-          'colsample_bytree': 1.0,
-          'num_threads' : -1,
-         }
+train_fc = save_path/f"train_oof_fc_v003_006.pkl"
+test_fc = save_path/f"test_oof_fc_v003_006.pkl"
 
-result_dict_lgb1 = train_model_regression(X=X,
-                                          X_test=X_test,
-                                          y=y_fc,
-                                          params=params,
-                                          folds=folds,
-                                          model_type='lgb',
-                                          eval_metric='group_mae',
-                                          plot_feature_importance=False,
-                                          verbose=500,
-                                          early_stopping_rounds=200,
-                                          n_estimators=8000)
-X['oof_fc'] = result_dict_lgb1['oof']
-X_test['oof_fc'] = result_dict_lgb1['prediction']
-to_pickle(submit_path/f"train_oof_fc_{DATA_VERSION}_{TRIAL_NO}.pkl", X['oof_fc'])
-to_pickle(submit_path/f"test_oof_fc_{DATA_VERSION}_{TRIAL_NO}.pkl", X_test['oof_fc'])
-to_pickle(model_path/f"first_model_list_{DATA_VERSION}_{TRIAL_NO}.pkl", result_dict_lgb1["models"])
+params = {'num_leaves': 128,
+    'min_child_samples': 79,
+    'objective': 'regression',
+    'max_depth': 9,
+    'learning_rate': 0.3,
+    "boosting_type": "gbdt",
+    "subsample_freq": 1,
+    "subsample": 0.9,
+    "bagging_seed": 11,
+    "metric": 'mae',
+    "verbosity": -1,
+    'reg_alpha': 0.1,
+    'reg_lambda': 0.3,
+    'colsample_bytree': 1.0,
+    'num_threads': -1,
+}
+
+if train_fc.exists() and test_fc.exists():
+    X['oof_fc'] = unpickle(train_fc)
+    X_test['oof_fc'] = unpickle(test_fc)
+else:
+    result_dict_lgb1 = train_model_regression(X=X,
+                                              X_test=X_test,
+                                              y=y_fc,
+                                              params=params,
+                                              folds=folds,
+                                              model_type='lgb',
+                                              eval_metric='group_mae',
+                                              plot_feature_importance=False,
+                                              verbose=500,
+                                              early_stopping_rounds=200,
+                                              n_estimators=6000)
+    X['oof_fc'] = result_dict_lgb1['oof']
+    X_test['oof_fc'] = result_dict_lgb1['prediction']
+    to_pickle(submit_path/f"train_oof_fc_{DATA_VERSION}_{TRIAL_NO}.pkl", X['oof_fc'])
+    to_pickle(submit_path/f"test_oof_fc_{DATA_VERSION}_{TRIAL_NO}.pkl", X_test['oof_fc'])
+    to_pickle(model_path/f"first_model_list_{DATA_VERSION}_{TRIAL_NO}.pkl", result_dict_lgb1["models"])
 
 X_short = pd.DataFrame({'ind': list(X.index), 'type': X['type'].values, 'oof': [0] * len(X), 'target': y.values})
 X_short_test = pd.DataFrame({'ind': list(X_test.index), 'type': X_test['type'].values, 'prediction': [0] * len(X_test)})
@@ -612,12 +628,12 @@ for t in X['type'].unique():
                                               y=y_t,
                                               params=params,
                                               folds=folds,
-                                              model_type='lgb',
+                                              model_type='xgb',
                                               eval_metric='group_mae',
                                               plot_feature_importance=True,
                                               verbose=500,
                                               early_stopping_rounds=200,
-                                              n_estimators=15000,
+                                              n_estimators=7000,
                                               mol_type=t)
     X_short.loc[X_short['type'] == t, 'oof'] = result_dict_lgb3['oof']
     X_short_test.loc[X_short_test['type'] == t, 'prediction'] = result_dict_lgb3['prediction']
@@ -626,6 +642,7 @@ for t in X['type'].unique():
     X_short_test.to_csv(submit_path/f"tmp_sub_{t}.csv")
     to_pickle(model_path/f"second_model_list_{DATA_VERSION}_{TRIAL_NO}.pkl", result_dict_lgb3["models"])
 
+sub = pd.read_csv(f'{file_folder}/sample_submission.csv')
 sub['scalar_coupling_constant'] = X_short_test['prediction']
 sub.to_csv(submit_path/f'submission_t_{DATA_VERSION}_{TRIAL_NO}.csv', index=False)
 print(sub.head())
